Fix bug when callback is called twice because of timer set twice or more times
There may be a situation when Lock#acquire() executes getData() and result is NO_NODE error which makes it call getChildren() and enter the same procedure again calling setTimer() twice or more times. In last mentioned function there was no check for already existing timer. That's the cause of the bug.